feat(notification): add auto user notification
This commit implements a feature that notifies the user when there is
update in their appointment. This is partial implementation of #5.

WIP: retrieval of user notification is to be implemented

NOTE: the cause for the weird behaviour in mentioned in  159d07b is an
ObjectID stored as string in the Database.